chore: move each wallet to its own file
